feat(analytics): refactor to hook
